# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'New India Export / VIRASTRA API'
  version: 2.3.0
  description: "Production Express API for customer onboarding, OTP auth, KYC, cases, Razorpay payments, GST invoices, events, and admin ops.\n\n**Auth:** Email OTP only. Send `Authorization: Bearer <JWT>` for protected routes.\n\n**Money:** Amounts in INR rupees in API/DB; converted to paise only for Razorpay.\n\n**Envelope:** `{ success, data?, message? }` on success; `{ success: false, message, code?, details? }` on errors.\n\nLegacy vault/shipment routes under `/api/cases/...` exist only when `ENABLE_LEGACY_SHIPMENT=true`."
  contact:
    name: 'New India Export'
    email: support@newindiaexport.com
servers:
  -
    url: 'http://127.0.0.1:5001'
    description: Local
  -
    url: /
    description: 'Current host'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT from OTP verify / login. Claims: sub, email, role, status.'
...
  schemas:
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data: {}
        message:
          type: string
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        code:
          type: string
        details:
          type: array
          items: {}
...
  responses:
    BadRequest:
      description: 'Validation / bad request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 'Missing or invalid JWT'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 'Insufficient role / ACL'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 'Resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /api/admin/analytics/overview:
    get:
      summary: 'Analytics overview KPIs'
      tags:
        - Admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
  /api/admin/analytics/summary:
    get:
      summary: 'Analytics summary'
      tags:
        - Admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
  /api/admin/analytics/detail:
    get:
      summary: 'Analytics detail series (placeholder)'
      tags:
        - Admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
  /api/admin/audit:
    get:
      summary: 'Audit log'
      tags:
        - Admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: limit
          in: query
          required: false
          schema:
            type: integer
        -
          name: action
          in: query
          required: false
          schema:
            type: string
        -
          name: actor
          in: query
          required: false
          schema:
            type: string
        -
          name: resourceId
          in: query
          required: false
          schema:
            type: string
  /api/admin/email-outbox:
    get:
      summary: 'Email outbox'
      tags:
        - Admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: status
          in: query
          required: false
          schema:
            type: string
  /api/admin/email-outbox/retry:
    post:
      summary: 'Retry failed emails'
      tags:
        - Admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                limit:
                  type: integer
  /api/admin/pending-counts:
    get:
      summary: 'Pending KYC + staff counts'
      tags:
        - Admin
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
...